Refactor files to be arranged by page
