+++ index.html
@@ -17,7 +17,7 @@
     <h1>Hello from Electron renderer!</h1>
     <p>👋</p>
     <p id="info"></p>
-    <a href="battle.html">
+    <a href="./src/pages/battle/battle.html">
         <div>
             <button type="button" name="startGame">Click here to start the game!</button>
         </div>
