+++ src/pages/battle/battle.html
@@ -11,7 +11,7 @@
       http-equiv="X-Content-Security-Policy"
       content="default-src 'self'; script-src 'self'"
     />
-    <link rel="stylesheet" href="./src/battle.css">
+    <link rel="stylesheet" href="./battle.css">
     <title>Daemon Test Fighter</title>
 
   </head>
@@ -66,6 +66,6 @@
       </div>
     </div>
     
-    <script type="module" src="./scripts/renderer-battle.js"></script>
+    <script type="module" src="./renderer-battle.js"></script>
   </body>
 </html>
