+++ src/pages/interlude/interlude.html
@@ -11,7 +11,7 @@
       http-equiv="X-Content-Security-Policy"
       content="default-src 'self'; script-src 'self'"
     />
-    <link rel="stylesheet" href="./src/interlude.css">
+    <link rel="stylesheet" href="./interlude.css">
     <title>Daemon Test Fighter</title>
 
   </head>
